moved datapoint related python tests into their own file
